Move mailer into its own package
Resolves #1
